+++ src/components/AboutProton/AboutProton.js
@@ -6,7 +6,7 @@ import IconIg from '../../pictures/icon-ig.svg'
 import IconMd from '../../pictures/icon-md.svg'
 import IconRd from '../../pictures/icon-rd.svg'
 import Mountain from '../../pictures/mountain.jpg'
-import MountainOverlay from '../../../public/mountainOverlay.jpg'
+import MountainOverlay from '../../pictures/mountain-overlay.jpg'
 
 export default function AboutProton() {
     return (
