+++ src/pages/index.js
@@ -4,10 +4,10 @@ import './index.scss'
 import WelcomeProton from '../components/WelcomeProton/WelcomeProton'
 import ProtonProducts from '../components/ProtonProducts/ProtonProducts'
 import AboutProton from '../components/AboutProton/AboutProton'
-import PictureMail from '../pictures/DP PM.png'
-import PictureVPN from '../pictures/DP PVPN.png'
-import PictureCalendar from '../pictures/DP PC.png'
-import PictureDrive from '../pictures/DP PD.png'
+import PictureMail from '../pictures/DP-PM.png'
+import PictureVPN from '../pictures/DP-PVPN.png'
+import PictureCalendar from '../pictures/DP-PC.png'
+import PictureDrive from '../pictures/DP-PD.png'
 import LogoMail from '../pictures/logo-mail.svg'
 import LogoCalendar from '../pictures/logo-cal.svg'
 import LogoDrive from '../pictures/logo-drive.svg'
